some changes

+++ backend/FileService/FileService.Core/FileStorage/IChunkSizeCalculator.cs
@@ -5,6 +5,6 @@ namespace FileService.Core.FileStorage;
 
 public interface IChunkSizeCalculator
 {
-    Result<(long ChunkSize, int TotalChunks), Error> CalculateChunkSize(
+    Result<(int ChunkSize, int TotalChunks), Error> CalculateChunkSize(
         long fileSiz);
 }
